--- .turbo/preferences/tui.json
@@ -1,4 +1,0 @@
-{
-  "is_task_list_visible": true,
-  "active_task": "app#dev"
-}
