+++ packages/web/i18n/en/account_team.json
@@ -244,12 +244,15 @@
   "search_test": "Search Test",
   "set_invoice_header": "Set up invoice header",
   "set_name_avatar": "Team avatar",
+  "skill.folder": "Folder",
+  "skill.skill": "Skill",
   "sync_immediately": "Synchronize now",
   "sync_member_failed": "Synchronization of members failed",
   "sync_member_success": "Synchronize members successfully",
   "total_team_members": "Total {{amount}} members",
   "transfer_app_ownership": "Transfer app ownership",
   "transfer_dataset_ownership": "Transfer dataset ownership",
+  "transfer_skill_ownership": "Transfer skill ownership",
   "transfer_ownership": "Transfer ownership",
   "transfer_team_ownership": "Transfer Team",
   "transfer_success": "Transfer successful",
@@ -275,11 +278,34 @@
   "update_data": "Update data",
   "update_dataset": "Update the knowledge base",
   "update_dataset_collaborator": "Knowledge Base Permission Changes",
+  "update_skill": "Update skill",
+  "update_skill_collaborator": "Skill permission change",
   "update_publish_app": "Application update",
   "used_times_limit": "Limit",
   "user_name": "username",
   "user_team_invite_member": "Invite members",
   "user_team_leave_team": "Leave the team",
   "user_team_leave_team_failed": "Failure to leave the team",
-  "waiting": "To be accepted"
+  "waiting": "To be accepted",
+  "copy_skill": "Copy skill",
+  "create_skill": "Create skill",
+  "create_skill_folder": "Create skill folder",
+  "delete_skill": "Delete skill",
+  "delete_skill_collaborator": "Skill permission deletion",
+  "deploy_skill": "Deploy skill",
+  "export_skill": "Export skill",
+  "import_skill": "Import skill",
+  "log_copy_skill": "[{{name}}] Copied [{{skillType}}] named [{{skillName}}]",
+  "log_create_skill": "[{{name}}] Created [{{skillType}}] named [{{skillName}}]",
+  "log_create_skill_folder": "[{{name}}] Created a skill folder named [{{folderName}}]",
+  "log_delete_skill": "[{{name}}] Deleted [{{skillType}}] named [{{skillName}}]",
+  "log_delete_skill_collaborator": "[{{name}}] Deleted the [{{itemName}}] permission named [{{itemValueName}}] from [{{skillType}}] named [{{skillName}}]",
+  "log_deploy_skill": "[{{name}}] Deployed [{{skillType}}] named [{{skillName}}]",
+  "log_export_skill": "[{{name}}] Exported [{{skillType}}] named [{{skillName}}]",
+  "log_import_skill": "[{{name}}] Imported [{{skillType}}] named [{{skillName}}]",
+  "log_move_skill": "[{{name}}] Moved [{{skillType}}] named [{{skillName}}] to [{{targetFolderName}}]",
+  "log_transfer_skill_ownership": "[{{name}}] Transferred ownership of [{{skillType}}] named [{{skillName}}] from [{{oldOwnerName}}] to [{{newOwnerName}}]",
+  "log_update_skill": "[{{name}}] Updated [{{skillType}}] named [{{skillName}}]",
+  "log_update_skill_collaborator": "[{{name}}] Updated collaborators for [{{skillType}}] named [{{skillName}}] to: Organization: [{{orgList}}], Group: [{{groupList}}], Member [{{tmbList}}]; permissions updated to: [{{permission}}]",
+  "move_skill": "Move skill"
 }
