+++ packages/web/i18n/zh-CN/account_team.json
@@ -244,12 +244,15 @@
   "search_test": "搜索测试",
   "set_invoice_header": "设置发票抬头",
   "set_name_avatar": "团队头像 & 团队名",
+  "skill.folder": "文件夹",
+  "skill.skill": "技能",
   "sync_immediately": "立即同步",
   "sync_member_failed": "同步成员失败",
   "sync_member_success": "同步成员成功",
   "total_team_members": "共 {{amount}} 名成员",
   "transfer_app_ownership": "转移应用所有权",
   "transfer_dataset_ownership": "转移知识库所有权",
+  "transfer_skill_ownership": "转移技能所有权",
   "transfer_ownership": "转让所有者",
   "transfer_team_ownership": "转让团队",
   "transfer_success": "转让成功",
@@ -275,11 +278,34 @@
   "update_data": "更新数据",
   "update_dataset": "更新知识库",
   "update_dataset_collaborator": "知识库权限更改",
+  "update_skill": "更新技能",
+  "update_skill_collaborator": "技能权限更改",
   "update_publish_app": "应用更新",
   "used_times_limit": "有效人数",
   "user_name": "用户名",
   "user_team_invite_member": "邀请成员",
   "user_team_leave_team": "离开团队",
   "user_team_leave_team_failed": "离开团队失败",
-  "waiting": "待接受"
+  "waiting": "待接受",
+  "copy_skill": "复制技能",
+  "create_skill": "创建技能",
+  "create_skill_folder": "创建技能文件夹",
+  "delete_skill": "删除技能",
+  "delete_skill_collaborator": "技能权限删除",
+  "deploy_skill": "部署技能",
+  "export_skill": "导出技能",
+  "import_skill": "导入技能",
+  "log_copy_skill": "【{{name}}】复制了名为【{{skillName}}】的【{{skillType}}】",
+  "log_create_skill": "【{{name}}】创建了名为【{{skillName}}】的【{{skillType}}】",
+  "log_create_skill_folder": "【{{name}}】创建了名为【{{folderName}}】的技能文件夹",
+  "log_delete_skill": "【{{name}}】删除了名为【{{skillName}}】的【{{skillType}}】",
+  "log_delete_skill_collaborator": "【{{name}}】将名为【{{skillName}}】的【{{skillType}}】中名为【{{itemValueName}}】的【{{itemName}}】权限删除",
+  "log_deploy_skill": "【{{name}}】部署了名为【{{skillName}}】的【{{skillType}}】",
+  "log_export_skill": "【{{name}}】导出了名为【{{skillName}}】的【{{skillType}}】",
+  "log_import_skill": "【{{name}}】导入了名为【{{skillName}}】的【{{skillType}}】",
+  "log_move_skill": "【{{name}}】将名为【{{skillName}}】的【{{skillType}}】移动到【{{targetFolderName}}】",
+  "log_transfer_skill_ownership": "【{{name}}】将名为【{{skillName}}】的【{{skillType}}】的所有权从【{{oldOwnerName}}】转移到【{{newOwnerName}}】",
+  "log_update_skill": "【{{name}}】更新了名为【{{skillName}}】的【{{skillType}}】",
+  "log_update_skill_collaborator": "【{{name}}】将名为【{{skillName}}】的【{{skillType}}】的合作者更新为：组织：【{{orgList}}】，群组：【{{groupList}}】，成员【{{tmbList}}】；权限更新为：【{{permission}}】",
+  "move_skill": "移动技能"
 }
