+++ pnpm-workspace.yaml
@@ -1,46 +1,46 @@
 packages:
   - packages/*
   - projects/*
-  - pro/*
+  - pro/admin
+  - pro/sso
   - scripts/icon
   - sdk/*
 
 overrides:
-  '@types/react': ^18
-  '@types/react-dom': ^18
+  "@types/react": ^18
+  "@types/react-dom": ^18
   react: ^18
   react-dom: ^18
 
 catalog:
-  '@chakra-ui/anatomy': ^2
-  '@chakra-ui/icons': ^2
-  '@chakra-ui/next-js': ^2
-  '@chakra-ui/react': ^2
-  '@chakra-ui/styled-system': ^2
-  '@chakra-ui/system': ^2
-  '@emotion/react': ^11
-  '@emotion/styled': ^11
-  '@modelcontextprotocol/sdk': ^1
-  '@fastgpt-sdk/storage': 0.6.15
-  '@fastgpt-sdk/logger': 0.1.2
-  '@types/lodash': ^4
-  '@types/react': ^18
-  '@types/react-dom': ^18
-  '@types/node': ^20
+  "@chakra-ui/anatomy": ^2
+  "@chakra-ui/icons": ^2
+  "@chakra-ui/next-js": ^2
+  "@chakra-ui/react": ^2
+  "@chakra-ui/styled-system": ^2
+  "@chakra-ui/system": ^2
+  "@emotion/react": ^11
+  "@emotion/styled": ^11
+  "@modelcontextprotocol/sdk": ^1
+  "@fastgpt-sdk/storage": 0.6.15
+  "@fastgpt-sdk/logger": 0.1.2
+  "@types/lodash": ^4
+  "@types/react": ^18
+  "@types/react-dom": ^18
+  "@types/node": ^20
   axios: 1.13.6
   date-fns: ^3
   dayjs: 1.11.19
   eslint: ^8
-  eslint-config-next: 15.5.12
+  eslint-config-next: 16.2.4
   express: ^4
   i18next: 23.16.8
   js-yaml: ^4.1.1
   json5: ^2.2.3
   lodash: 4.17.23
   minio: 8.0.7
-  next: 16.1.6
+  next: 16.2.4
   next-i18next: 15.4.2
-  next-rspack: 16.1.6
   proxy-agent: ^6
   react: ^18
   react-dom: ^18
