+++ projects/app/src/pages/api/core/dataset/collection/paths.ts
@@ -41,7 +41,7 @@ export async function getDatasetCollectionPaths({
 
     if (!parent) return [];
 
-    const paths = await find(parent.parentId);
+    const paths = await find(parent.parentId ?? undefined);
     paths.push({ parentId, parentName: parent.name });
 
     return paths;
