+++ projects/app/src/web/core/dataset/constants.ts
@@ -2,6 +2,7 @@ import { defaultQAModels, defaultVectorModels } from '@fastgpt/global/core/ai/co
 import {
   DatasetCollectionDataProcessModeEnum,
   DatasetCollectionTypeEnum,
+  DatasetStatusEnum,
   DatasetTypeEnum
 } from '@fastgpt/global/core/dataset/constants';
 import type { DatasetCollectionItemType, DatasetItemType } from '@fastgpt/global/core/dataset/type';
@@ -19,7 +20,7 @@ export const defaultDatasetDetail: DatasetItemType = {
   avatar: '/icon/logo.svg',
   name: '',
   intro: '',
-  status: 'active',
+  status: DatasetStatusEnum.active,
   permission: new DatasetPermission(),
   vectorModel: defaultVectorModels[0],
   agentModel: defaultQAModels[0],
